requirements.txt
Flask 
pyswisseph
pyzt